Clean architecture implemented on project
