+++ analysis_options.yaml
@@ -22,6 +22,14 @@ linter:
   # `// ignore_for_file: name_of_lint` syntax on the line or in the file
   # producing the lint.
   rules:
+    camel_case_types: true
+    prefer_relative_imports: true
+    require_trailing_commas: true
+    package_names: true
+    eol_at_end_of_file: true
+    always_declare_return_types: true
+
+
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
