--- lib/custom_text_widget.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/material.dart';
-import '/constants.dart';
-
-class CustomText extends StatelessWidget {
-  final String text;
-  final double fontSize;
-  final int color;
-  static const double fontSizeDefaultValue = 12;
-
-  const CustomText({
-    required this.text,
-    this.fontSize = fontSizeDefaultValue,
-    this.color = AppConstants.appFontColor,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      text,
-      style: TextStyle(
-        color: Color(
-          color,
-        ),
-        fontSize: fontSize,
-      ),
-    );
-  }
-}
