--- lib/model/genre_data_model.dart
@@ -1,9 +1,0 @@
-class GenreDataModel {
-  num id;
-  String name;
-
-  GenreDataModel({
-    required this.id,
-    required this.name,
-  });
-}
