+++ lib/src/domain/entity/genre.dart
@@ -1,0 +1,9 @@
+class Genre {
+  num id;
+  String name;
+
+  Genre({
+    required this.id,
+    required this.name,
+  });
+}
