+++ lib/src/presentation/widget/movie_poster.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import '../../core/util/constants.dart';
+
+class MoviePoster extends StatelessWidget {
+  String posterImageUrl;
+  double posterImageWidth;
+  double posterImageHeight;
+  static const double paddingCircleProgressIndicator = 16;
+
+  MoviePoster({
+    required this.posterImageUrl,
+    required this.posterImageWidth,
+    required this.posterImageHeight,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.network(
+      posterImageUrl,
+      width: posterImageWidth,
+      height: posterImageHeight,
+      loadingBuilder: (
+        BuildContext context,
+        Widget child,
+        ImageChunkEvent? loadingProgress,
+      ) {
+        if (loadingProgress != null) {
+          return const Padding(
+            padding: EdgeInsets.all(
+              paddingCircleProgressIndicator,
+            ),
+            child: CircularProgressIndicator(
+              color: Color(
+                AppConstants.appItemsBackgroundColor,
+              ),
+            ),
+          );
+        } else {
+          return child;
+        }
+      },
+    );
+  }
+}
